+++ pytest/test_misk.py
@@ -1,0 +1,5 @@
+from utils import *
+
+def test_toppage(client):
+    resp = client.get('/')
+    assert b'DOC' in resp.data
